+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - extra default task future
 - skipping default screens in dev mode
 - update settings menu
+- add setting to choose startup tab
+- add notification quiet hours setting
 
 ## [0.1.55] - 2026-02-26
 - recurring tasks